requirements.txt
psycopg2
pytest
requests
solrpy
